feat: App, DashBoard 컴포넌트 분리 리팩토링 - AppContent, AuthHeader, SplashScreen, LoadingScreen, ProtectedRoute 컴포넌트를 새로 추가하여 코드 구조 개선 - 인증 상태 및 네비게이션 로직을 커스텀 훅으로 분리하여 재사용성 향상 - 대시보드 및 로그인 관련 UI 구성 요소를 통합하여 사용자 경험 개선
