+++ src/components/layout/dashboard/index.ts
@@ -2,4 +2,6 @@
  * 대시보드 관련 컴포넌트 통합 export
  */
 export { default as Dashboard } from './Dashboard';
-export { default as SocialAccountLink } from './SocialAccountLink';
+
+// Dashboard 섹션 컴포넌트들
+export * from './sections';
