+++ src/components/layout/dashboard/sections/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Dashboard 섹션 컴포넌트 통합 export
+ */
+export { default as UserInfoSection } from './UserInfoSection';
+export { default as TokenInfoSection } from './TokenInfoSection';
+export { default as AutoTokenRefreshSection } from './AutoTokenRefreshSection';
+export { default as TokenManagementSection } from './TokenManagementSection';
+export { default as SocialAccountLink } from './SocialAccountLink';
+
