+++ src/components/layout/index.ts
@@ -1,4 +1,12 @@
 /**
  * 레이아웃 관련 컴포넌트 통합 export
  */
+export { default as AppContent } from './AppContent';
+export { default as SplashScreen } from './SplashScreen';
+export { default as LoadingScreen } from './LoadingScreen';
+export { default as ProtectedRoute } from './ProtectedRoute';
+export { default as AuthHeader } from './AuthHeader';
 export { default as BottomSheet } from './BottomSheet';
+
+// Dashboard 관련 컴포넌트들
+export * from './dashboard';
